- add git-ignore; - correct namespaces in Device project; - fix list selectcommand handling;

+++ Cinegy.ImportTool.Infrastructure/Model/ITrack.cs
@@ -4,7 +4,7 @@
     {
         #region Properties
 
-        int Length { get; }
+        long Length { get; }
         string Name { get; }
         TrackType Type { get; }
